+++ tests/test_futureproof.py
@@ -35,9 +35,8 @@ def flaky_sum(a, b, delay=0):
     return a + b
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_raise_immediate_exceptions(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_raise_immediate_exceptions():
+    executor = conftest.get_executor_for_type()
     tm = futureproof.TaskManager(executor)
 
     tm.submit(custom_sum, 1)
@@ -49,9 +48,8 @@ def test_raise_immediate_exceptions(executor_type):
     )
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_log_immediate_exceptions(executor_type, mocker):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_log_immediate_exceptions(mocker):
+    executor = conftest.get_executor_for_type()
     mock_logger = mocker.patch("futureproof.task_manager.logger")
     tm = futureproof.TaskManager(executor, error_policy=futureproof.ErrorPolicyEnum.LOG)
 
@@ -61,9 +59,8 @@ def test_log_immediate_exceptions(executor_type, mocker):
     assert mock_logger.exception.call_count == 1
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_context_manager_raise_immediate_exceptions(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_context_manager_raise_immediate_exceptions():
+    executor = conftest.get_executor_for_type()
     with pytest.raises(TypeError) as exc_info:
         with futureproof.TaskManager(executor) as tm:
             tm.submit(custom_sum, 1)
@@ -73,9 +70,8 @@ def test_context_manager_raise_immediate_exceptions(executor_type):
     )
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_context_manager_log_immediate_exceptions(executor_type, mocker):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_context_manager_log_immediate_exceptions(mocker):
+    executor = conftest.get_executor_for_type()
     mock_logger = mocker.patch("futureproof.task_manager.logger")
 
     with futureproof.TaskManager(
@@ -87,10 +83,9 @@ def test_context_manager_log_immediate_exceptions(executor_type, mocker):
 
 
 @pytest.mark.timeout(60)
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.slow
-def test_submit_valid_functions(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_submit_valid_functions():
+    executor = conftest.get_executor_for_type()
     tm = futureproof.TaskManager(executor)
 
     for i in range(100):
@@ -100,9 +95,8 @@ def test_submit_valid_functions(executor_type):
     assert list(range(1, 101)) == sorted(tm.results)
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_submit_flaky_functions(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_submit_flaky_functions():
+    executor = conftest.get_executor_for_type()
     tm = futureproof.TaskManager(executor)
 
     for i in range(1, 101):
@@ -112,7 +106,9 @@ def test_submit_flaky_functions(executor_type):
         tm.run()
 
     # processes tend to take longer to shutdown, threads are faster
-    min_results, max_results = (5, 6) if executor_type == "threads" else (5, 10)
+    min_results, max_results = (
+        (5, 6) if isinstance(executor, futureproof.ThreadPoolExecutor) else (5, 10)
+    )
     assert min_results <= len(tm.results) < max_results
     failed_task = next(
         task for task in tm.completed_tasks if isinstance(task.result, Exception)
@@ -120,16 +116,17 @@ def test_submit_flaky_functions(executor_type):
     assert isinstance(failed_task.result, ValueError)
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
-def test_submit_flaky_functions_context_manager(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_submit_flaky_functions_context_manager():
+    executor = conftest.get_executor_for_type()
     with pytest.raises(ValueError):
         with futureproof.TaskManager(executor) as tm:
             for i in range(1, 101):
                 tm.submit(flaky_sum, i, 1)
 
     # processes tend to take longer to shutdown, threads are faster
-    min_results, max_results = (5, 6) if executor_type == "threads" else (5, 10)
+    min_results, max_results = (
+        (5, 6) if isinstance(executor, futureproof.ThreadPoolExecutor) else (5, 10)
+    )
     assert min_results <= len(tm.results) < max_results
     failed_task = next(
         task for task in tm.completed_tasks if isinstance(task.result, Exception)
@@ -138,10 +135,9 @@ def test_submit_flaky_functions_context_manager(executor_type):
 
 
 @pytest.mark.timeout(60)
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.slow
-def test_map_generator(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_map_generator():
+    executor = conftest.get_executor_for_type()
     fn = partial(custom_sum, b=1)
     with futureproof.TaskManager(executor) as tm:
         tm.map(fn, range(100))
@@ -150,10 +146,9 @@ def test_map_generator(executor_type):
 
 
 @pytest.mark.timeout(60)
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.slow
-def test_map_lazy_generator(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_map_lazy_generator():
+    executor = conftest.get_executor_for_type()
 
     def gen():
         for i in range(100):
@@ -168,10 +163,9 @@ def test_map_lazy_generator(executor_type):
     assert list(range(1, 101)) == sorted(tm.results)
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.timeout(4)
-def test_submit_after_map(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_submit_after_map():
+    executor = conftest.get_executor_for_type()
     fn = partial(custom_sum, b=1)
     tm = futureproof.TaskManager(executor)
     tm.map(fn, range(9))
@@ -181,10 +175,9 @@ def test_submit_after_map(executor_type):
     assert list(range(1, 11)) == sorted(tm.results)
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.timeout(4)
-def test_map_after_submit(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_map_after_submit():
+    executor = conftest.get_executor_for_type()
     fn = partial(custom_sum, b=1)
     tm = futureproof.TaskManager(executor)
     tm.submit(fn, 0)
@@ -194,10 +187,9 @@ def test_map_after_submit(executor_type):
     assert list(range(1, 11)) == sorted(tm.results)
 
 
-@pytest.mark.parametrize("executor_type", ("thread", "process"))
 @pytest.mark.timeout(3)
-def test_as_completed(executor_type):
-    executor = conftest.get_executor_for_type(executor_type)
+def test_as_completed():
+    executor = conftest.get_executor_for_type()
     fn = partial(custom_sum, b=1)
     tm = futureproof.TaskManager(executor)
 
@@ -213,7 +205,17 @@ def test_as_completed(executor_type):
 
 
 def test_task_manager_error_policy_as_string():
-    executor = conftest.get_executor_for_type("thread")
+    executor = conftest.get_executor_for_type()
     tm = futureproof.TaskManager(executor, "log")
 
     assert tm._error_policy == futureproof.ErrorPolicyEnum.LOG
+
+
+def test_process_pool_executor_3_7_only(mocker):
+    mock_version = mocker.patch("futureproof.executors.sys")
+    mock_version.version_info = (3, 6, 6)
+    with pytest.raises(NotImplementedError):
+        futureproof.ProcessPoolExecutor()
+
+    mock_version.version_info = (3, 7, 0)
+    assert futureproof.ProcessPoolExecutor()